+++ src/visuals-api.d.ts
@@ -416,6 +416,9 @@ declare module powerbi {
 
     export interface DataRolesInfo {
         drillableRoles?: powerbi.DrillableRoles;
+
+        /** Indicates whether the drill is disabled */
+        isDrillDisabled?: boolean;
     }
 
     export interface DataViewMetadataColumn {
@@ -1530,21 +1533,42 @@ declare module powerbi.extensibility {
 }
 
 declare module powerbi.extensibility {
-    export interface AcquireAADTokenResult {
-        accessToken?: string;
+
+    /**
+     * Interface representing information about the user associated with the token.
+     */
+    export interface AcquireAADTokenUserInfo {
+        userId?: string;   // Unique identifier for the user
+        tenantId?: string; // Unique identifier for the tenant
     }
 
+    /**
+     * Interface representing the result of acquiring a Microsoft Entra ID token.
+     */
+    export interface AcquireAADTokenResult {
+        accessToken?: string;       // Access token issued by Microsoft Entra ID
+        expiresOn?: number;         // Expiration time of the access token
+        userInfo?: AcquireAADTokenUserInfo;     // Information about the user associated with the token
+    }
+
+    /**
+     * Interface representing a service for acquiring authentication tokens from Microsoft Entra ID.
+     */
     export interface IAcquireAADTokenService {
-        /** Returns an authentication token for the resource that the visual defined as a privilge
-         * and the scope is the visual guid plus a constant string "_CV_ForPBI"
-         * @returns the promise that resolves to the authentication token
-        */
+        /** 
+         * Retrieves an authentication token payload.
+         * 
+         * The audience is determined by the visual's `AADAuthentication` privilege parameter. 
+         * The scope is formed by concatenating the visual's GUID with "_CV_ForPBI".
+         * 
+         * @returns A promise that resolves to the authentication token payload.
+         */
         acquireAADToken(): IPromise<AcquireAADTokenResult>;
 
         /**
-         * Returns the availability status of the service.
+         * Retrieves the availability status of the service.
          * 
-         * @returns the promise that resolves to privilege status of the service
+         * @returns A promise that resolves to the privilege status of the service.
          */
         acquireAADTokenstatus(): IPromise<PrivilegeStatus>;
     }
@@ -1611,50 +1635,6 @@ declare module powerbi.extensibility {
         /**
          * Deletes data associated with 'key' from local storage.
          * 
-         * @param key - the name of the payload to remove
-         */
-        remove(key: string): void;
-    }
-}
-
-declare module powerbi.extensibility {
-
-    interface StorageV2ResultInfo {
-        success: boolean;
-    }
-
-    /**
-     * Provides an access to local storage for read / write access
-     */
-    interface IVisualLocalStorageV2Service {
-        /**
-         * Returns the availability status of the service.
-         *
-         * @returns the promise that resolves to privilege status of the service
-         */
-        status(): IPromise<PrivilegeStatus>;
-
-        /**
-         * Returns promise that resolves to the data associated with 'key' if it was found or rejects otherwise.
-         *
-         * @param key - the name of the payload to retrieve
-         * @returns the promise that resolves to the data required or rejects if it wasn't found or an error occured.
-         */
-        get(key: string): IPromise<string>;
-
-        /**
-         * Saves the data to local storage. This data can be later be retrieved using the 'key'.
-         * Returns a promise that resolves to StorageV2ResultInfo, or rejects if an error occured.
-         *
-         * @param key - the name of the payload to store
-         * @param data - the payload string to store
-         * @returns the promise resolves to StorageV2ResultInfo, or rejects if an error occured.
-         */
-        set(key: string, data: string): IPromise<StorageV2ResultInfo>;
-
-        /**
-         * Deletes data associated with 'key' from local storage.
-         *
          * @param key - the name of the payload to remove
          */
         remove(key: string): void;
@@ -1732,6 +1712,58 @@ declare module powerbi.extensibility {
     }
 }
 
+declare module powerbi.extensibility {
+    interface StorageV2ResultInfo {
+        success: boolean;
+    }
+
+    /**
+     * Provides an access to local storage for read / write access
+     */
+    interface IVisualLocalStorageV2Service {
+        /**
+         * Returns the availability status of the service.
+         *
+         * @returns the promise that resolves to privilege status of the service
+         */
+        status(): IPromise<PrivilegeStatus>;
+
+        /**
+         * Returns promise that resolves to the data associated with 'key' if it was found or rejects otherwise.
+         *
+         * @param key - the name of the payload to retrieve
+         * @returns the promise that resolves to the data required or rejects if it wasn't found or an error occured.
+         */
+        get(key: string): IPromise<string>;
+
+        /**
+         * Saves the data to local storage. This data can be later be retrieved using the 'key'.
+         * Returns a promise that resolves to StorageV2ResultInfo, or rejects if an error occured.
+         *
+         * @param key - the name of the payload to store
+         * @param data - the payload string to store
+         * @returns the promise resolves to StorageV2ResultInfo, or rejects if an error occured.
+         */
+        set(key: string, data: string): IPromise<StorageV2ResultInfo>;
+
+        /**
+         * Deletes data associated with 'key' from local storage.
+         *
+         * @param key - the name of the payload to remove
+         */
+        remove(key: string): void;
+    }
+}
+
+declare module powerbi.extensibility {
+    export interface IVisualSubSelectionService {
+        //** Emits the custom visual's sub-selection to PowerBI */
+        subSelect(subSelection: powerbi.visuals.CustomVisualSubSelection): void;
+        //** Sends the custom visual's sub-selection outlines to the PowerBI's outline renderer */
+        updateRegionOutlines(outlines: powerbi.visuals.SubSelectionRegionOutline[]): void;
+    }
+}
+
 declare namespace powerbi.common {
     export const enum CustomVisualHostEnv {
         Web = 1 << 0,
@@ -1740,12 +1772,21 @@ declare namespace powerbi.common {
         Embed = 1 << 3,
         ReportServer = 1 << 4,
         ExportReportHost = 1 << 5,
-        Mobile = 1 << 6
+        Mobile = 1 << 6,
+        DashboardHost = 1 << 7
     }
 }
 
 declare module powerbi {
     export interface IFilter { }
+}
+
+declare module powerbi.extensibility {
+    /** Provides encapsulated utility functions for the visual. */
+    export interface ICustomVisualsOpaqueUtils {
+        /** Compares the two CustomVisualOpaqueIdentity values for equality. */
+        compareCustomVisualOpaqueIdentities(identity1: powerbi.visuals.CustomVisualOpaqueIdentity, identity2: powerbi.visuals.CustomVisualOpaqueIdentity): boolean;
+    }
 }
 
 /**
@@ -1768,6 +1809,9 @@ declare module powerbi.extensibility.visual {
 
         /** Gets the settings to display in the formatting pane */
         getFormattingModel?(): visuals.FormattingModel | undefined;
+
+        //** Contains On Object get APIs */
+        visualOnObjectFormatting?: VisualOnObjectFormatting;
     }
 
     export interface IVisualHost extends extensibility.IVisualHost {
@@ -1797,8 +1841,17 @@ declare module powerbi.extensibility.visual {
         webAccessService: IWebAccessService;
         drill: (args: DrillArgs) => void;
         applyCustomSort: (args: CustomVisualApplyCustomSortArgs) => void;
-        storageV2Service: IVisualLocalStorageV2Service;
         acquireAADTokenService: IAcquireAADTokenService;
+        setCanDrill: (drillAllowed: boolean) => void;
+        storageV2Service: IVisualLocalStorageV2Service;
+        subSelectionService: IVisualSubSelectionService;
+        createOpaqueUtils: () => ICustomVisualsOpaqueUtils;
+    }
+
+    export interface VisualOnObjectFormatting {
+        getSubSelectionStyles(subSelections: powerbi.visuals.CustomVisualSubSelection[]): powerbi.visuals.SubSelectionStyles | undefined;
+        getSubSelectionShortcuts(subSelections: powerbi.visuals.CustomVisualSubSelection[], filter: powerbi.visuals.SubSelectionShortcutsKey | undefined): powerbi.visuals.VisualSubSelectionShortcuts | undefined;
+        getSubSelectables?(filter?: powerbi.visuals.SubSelectionStylesType): powerbi.visuals.CustomVisualSubSelection[] | undefined;
     }
 
     export interface VisualUpdateOptions extends extensibility.VisualUpdateOptions {
@@ -1810,6 +1863,8 @@ declare module powerbi.extensibility.visual {
         operationKind?: VisualDataChangeOperationKind;
         jsonFilters?: IFilter[];
         isInFocus?: boolean;
+        subSelections?: powerbi.visuals.CustomVisualSubSelection[];
+        formatMode?: boolean;
     }
 
     export interface VisualConstructorOptions extends extensibility.VisualConstructorOptions {