+++ CHANGELOG.md
@@ -1,7 +1,23 @@
 # Change Log - Power BI Custom Visuals API
-## 5.6.0
-* Adds the `storageV2` service.
-* Adds the `acquireAADToken` service.
+## 5.9.0
+* `AcquireAADTokenResult` : Extended with additional properties.
+* `createOpaqueUtils` : Create an encapsulated utility for the visual.
+* Filter API : Exposing new filter type for hierarchy data - HierarchyIdentity.
+* `areHierarchicallyRelated` : Requires that the items in the role are hierarchically related.
+* `CustomVisualHostEnv`: Extended with `DashboardHost` type for representing tiles and dashboards in embedded environments. 
+
+## 5.8.0
+* `storageV2Service` : Enables visuals to use the browser's local storage.
+* Support For OnObject Formatting:
+    - Adds OnObject Formatting interfaces.
+    - visualOnObjectFormatting: Contains the get APIs of the onObject.
+    - `subSelectionService`: Enables visuals to send subSelections and outlines to PowerBI.
+    - Adds 2 capabilities: `supportsOnObjectFormatting` and `enablePointerEventsFormatMode`.
+    - Adds `subSelections` and  `formatMode` in `VisualUpdateOptions` interface.
+
+## 5.7.0
+* `acquireAADTokenService` : Enables visuals to obtain Microsoft Entra ID (formerly known as Azure AD) access tokens for signed-in users.
+* Dynamic drill control : Adds new API and capability to dynamically control the drill and expand collapse features.
 
 ## 5.4.0
 * Adds `isDataFilterApplied` into DataViewMetadata, to provide a boolean value of whether any applied filter affects the visual.